feat: separate handlers and flask app initiator. also add logging
